requirements.txt
selenium==4.10.0
mysql-connector-python==8.0.33
webdriver-manager==4.0.2
fake_useragent==1.1.1
textblob==0.15.3
